backend/requirements.txt
agno==2.3.24
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
apscheduler==3.11.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
email-validator==2.3.0
fastapi==0.104.1
fastapi-cli==0.0.20
fastapi-cloud-cli==0.9.0
fastar==0.8.0
gitdb==4.0.12
gitpython==3.1.46
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jinja2==3.1.6
jiter==0.12.0
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
openai==2.15.0
packaging==25.0
pydantic==2.5.0
pydantic-core==2.14.1
pydantic-extra-types==2.11.0
pydantic-settings==2.1.0
pygments==2.19.2
pymongo==4.16.0
python-dotenv==1.0.0
python-multipart==0.0.6
pyyaml==6.0.3
requests==2.31.0
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
sentry-sdk==2.49.0
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
starlette==0.27.0
tqdm==4.67.1
typer==0.21.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
